+++ src/pages/enhancement/ApprovalTable.js
@@ -4,15 +4,14 @@ import Typography from '@material-ui/core/Typography';
 import CloudDownloadIcon from '@material-ui/icons/CloudDownload';
 import { makeStyles } from '@material-ui/styles';
 import clsx from 'clsx';
-import moment from 'moment';
 import MUIDataTable from 'mui-datatables';
 import React, { useMemo, useState, useEffect } from 'react';
-import { useQuery } from 'react-query';
 import { NavLink as RouterLink } from 'react-router-dom';
-import MuiTableFooter from '../../../components/CommonComponents/MuiTableFooter';
-import Currency from '../../../components/Number/Currency';
-import { downloadRenewalData, getPageDetails, getRenewalLoanByStatus } from '../../../services/renewal.service';
-import { dateCustomSort } from '../../../utils/commonFunctions.util';
+import MuiTableFooter from '../../components/CommonComponents/MuiTableFooter';
+import Currency from '../../components/Number/Currency';
+import { downloadEnhancementData, getEnhancedLoanByStatus, getPageDetails } from '../../services/enhancement.service';
+import { dateCustomSort } from '../../utils/commonFunctions.util';
+
 
 const useStyles = makeStyles(theme => ({
   title: {
@@ -30,21 +29,17 @@ const useStyles = makeStyles(theme => ({
 }));
 
 
-const ReviewTable = ({ title, onRowClick, filterQry, currentUser }) => {
+const ApprovalTable = ({ title, onRowClick, filterQry, currentUser }) => {
   const classes = useStyles();
   const [loans, setLoans] = useState([]);
   const [page, setPage] = useState();
+  const [pageData, setPageData] = useState();
   const [search, setSearch] = useState();
   const [loading, setLoading] = useState(false);
 
-  const pageDetailsQuery = useQuery({
-    queryKey: ['renewal_approvalRecordCount', filterQry, search],
-    queryFn: () => getPageDetails('approval', filterQry),
-  });
-
   useEffect(() => {
     setLoading(true);
-    getRenewalLoanByStatus('approval', filterQry, page, search)
+    getEnhancedLoanByStatus('approval', filterQry, page, search)
       .then(data => {
         setLoans(data);
         setLoading(false);
@@ -54,8 +49,17 @@ const ReviewTable = ({ title, onRowClick, filterQry, currentUser }) => {
       })
   }, [filterQry, page, search])
 
+  useEffect(() => {
+    getPageDetails('approval', filterQry)
+      .then((res) => {
+        setPageData(res)
+      })
+      .catch((e) => console.log('getPageCountError >>>', e))
+  }, [filterQry])
+
+
   const onDownloadClick = () => {
-    downloadRenewalData('draft', filterQry)
+    downloadEnhancementData('approval', filterQry)
       .then(data => {
         window.open(data[0]?.url, '_blank')
       })
@@ -87,8 +91,18 @@ const ReviewTable = ({ title, onRowClick, filterQry, currentUser }) => {
         }
       },
       {
-        label: 'Type',
-        name: 'product_name',
+        label: 'Old Product Type',
+        name: 'old_product_name',
+        options: {
+          filter: false,
+          sort: true,
+          customBodyRender: value => <span className={clsx(classes.pill, classes[`pills_${value}`])}>{value}</span>
+        }
+      },
+      
+      {
+        label: 'New Product Type',
+        name: 'new_product_name',
         options: {
           filter: false,
           sort: true,
@@ -97,7 +111,7 @@ const ReviewTable = ({ title, onRowClick, filterQry, currentUser }) => {
       },
       {
         label: 'Region',
-        name: 'region_name',
+        name: 'region',
         options: {
           filter: false,
           sort: true,
@@ -106,41 +120,33 @@ const ReviewTable = ({ title, onRowClick, filterQry, currentUser }) => {
 
       },
       {
-        label: 'Req. Amount',
-        name: 'requested_amount',
+        label: 'Old loan Amount',
+        name: 'old_loan_amount',
         options: {
           filter: false,
           sort: true,
           setCellProps: () => ({
-            align: 'left',
+            align: 'right',
+          }),
+          setCellHeaderProps: () => ({
+            align: 'right',
           }),
           customBodyRender: value => <strong><Currency value={value} /></strong>
         }
       },
       {
-        label: 'Month of renewal',
-        name: 'renewal_month',
-        options: {
-          filter: true,
-          filterWidth: '100%',
-          sort: true,
-          setCellProps: () => ({
-            align: 'center',
-          }),
-          customBodyRender: value => {
-            return <div>{value ? moment(new Date(value), 'YYYY-MM-DD').format('MMM, YY') : '-'}</div>
-          }
-        }
-      },
-      {
-        label: 'Renewal Fee status',
-        name: 'renewal_fee_payment_status',
+        label: 'New Loan Amount',
+        name: 'new_loan_amount',
         options: {
           filter: false,
           sort: true,
-          customBodyRender: (value) => {
-            return <>{value?.toUpperCase()}</>
-          },
+          setCellProps: () => ({
+            align: 'right',
+          }),
+          setCellHeaderProps: () => ({
+            align: 'right',
+          }),
+          customBodyRender: value => <strong><Currency value={value} /></strong>
         }
       },
     ]
@@ -154,16 +160,26 @@ const ReviewTable = ({ title, onRowClick, filterQry, currentUser }) => {
     filter: false,
     print: false,
     sort: false,
+    download: false,
     viewColumns: false,
     searchPlaceholder: 'Search by dealreship ID/Name',
     onSearchChange: (searchText) => {
       setSearch(searchText)
     },
+    customToolbar: () => {
+      return (
+        <>
+          <Tooltip title="Download">
+            <Button style={{ marginTop: 0 }} size='small' startIcon={<CloudDownloadIcon style={{ width: 24, height: 24, color: '#525252' }} color="#f5f5f5" />} onClick={onDownloadClick}></Button>
+          </Tooltip>
+        </>
+      );
+    },
     customFooter: () => {
       return (
         <div style={{ display: 'flex', justifyContent: 'flex-end' }}>
           <MuiTableFooter
-            totalCount={pageDetailsQuery?.data?.total_number_of_pages}
+            totalCount={pageData?.total_number_of_pages}
             pageSize={10}
             onPageChange={(value) => { setPage(value) }}
           />
@@ -178,23 +194,15 @@ const ReviewTable = ({ title, onRowClick, filterQry, currentUser }) => {
     customSort: (data, dataIndex, rowIndex) => {
       let dateIndex = 5
       return dateCustomSort(data, dataIndex, rowIndex, dateIndex)
-    },
-    customToolbar: () => {
-      return (
-        <>
-          <Tooltip title="Download">
-            <Button style={{ marginTop: 0 }} size='small' startIcon={<CloudDownloadIcon style={{ width: 24, height: 24, color: '#525252' }} color="#f5f5f5" />} onClick={onDownloadClick}></Button>
-          </Tooltip>
-        </>
-      );
-    },
+    }
   };
 
   return (
     <div className={classes.root}>
       <MUIDataTable
-        title={title ? <Typography className={classes.title} variant="h4" component="h4">{title}</Typography> : null}
+        title={title ? <Typography className={classes.title} variant="h4" component="h4">{title} ({loans.length})</Typography> : null}
         data={loans}
+        style={classes.tableStyle}
         columns={columns}
         options={options}
       />
@@ -205,4 +213,4 @@ const ReviewTable = ({ title, onRowClick, filterQry, currentUser }) => {
   )
 }
 
-export default ReviewTable;
+export default ApprovalTable;