+++ src/pages/enhancement/SubmittedTable.js
@@ -4,15 +4,14 @@ import Typography from '@material-ui/core/Typography';
 import CloudDownloadIcon from '@material-ui/icons/CloudDownload';
 import { makeStyles } from '@material-ui/styles';
 import clsx from 'clsx';
-import moment from 'moment';
 import MUIDataTable from 'mui-datatables';
 import React, { useMemo, useState, useEffect } from 'react';
-import { useQuery } from 'react-query';
 import { NavLink as RouterLink } from 'react-router-dom';
-import MuiTableFooter from '../../../components/CommonComponents/MuiTableFooter';
-import Currency from '../../../components/Number/Currency';
-import { downloadRenewalData, getPageDetails, getRenewalLoanByStatus } from '../../../services/renewal.service';
-import { dateCustomSort } from '../../../utils/commonFunctions.util';
+import MuiTableFooter from '../../components/CommonComponents/MuiTableFooter';
+import Currency from '../../components/Number/Currency';
+import { downloadEnhancementData, getEnhancedLoanByStatus, getPageDetails } from '../../services/enhancement.service';
+import { dateCustomSort } from '../../utils/commonFunctions.util';
+
 
 const useStyles = makeStyles(theme => ({
   title: {
@@ -30,21 +29,17 @@ const useStyles = makeStyles(theme => ({
 }));
 
 
-const ReviewTable = ({ title, onRowClick, filterQry, currentUser }) => {
+const SubmittedTable = ({ title, onRowClick, filterQry }) => {
   const classes = useStyles();
   const [loans, setLoans] = useState([]);
   const [page, setPage] = useState();
+  const [pageData, setPageData] = useState();
   const [search, setSearch] = useState();
   const [loading, setLoading] = useState(false);
 
-  const pageDetailsQuery = useQuery({
-    queryKey: ['renewal_approvalRecordCount', filterQry, search],
-    queryFn: () => getPageDetails('approval', filterQry),
-  });
-
   useEffect(() => {
     setLoading(true);
-    getRenewalLoanByStatus('approval', filterQry, page, search)
+    getEnhancedLoanByStatus('submit', filterQry, page, search)
       .then(data => {
         setLoans(data);
         setLoading(false);
@@ -54,8 +49,17 @@ const ReviewTable = ({ title, onRowClick, filterQry, currentUser }) => {
       })
   }, [filterQry, page, search])
 
+  useEffect(() => {
+    getPageDetails('submit', filterQry)
+      .then((res) => {
+        setPageData(res)
+      })
+      .catch((e) => console.log('getPageCountError >>>', e))
+  }, [filterQry])
+
+
   const onDownloadClick = () => {
-    downloadRenewalData('draft', filterQry)
+    downloadEnhancementData('submit', filterQry)
       .then(data => {
         window.open(data[0]?.url, '_blank')
       })
@@ -87,8 +91,8 @@ const ReviewTable = ({ title, onRowClick, filterQry, currentUser }) => {
         }
       },
       {
-        label: 'Type',
-        name: 'product_name',
+        label: 'Product Type',
+        name: 'new_product_name',
         options: {
           filter: false,
           sort: true,
@@ -97,7 +101,7 @@ const ReviewTable = ({ title, onRowClick, filterQry, currentUser }) => {
       },
       {
         label: 'Region',
-        name: 'region_name',
+        name: 'region',
         options: {
           filter: false,
           sort: true,
@@ -106,41 +110,18 @@ const ReviewTable = ({ title, onRowClick, filterQry, currentUser }) => {
 
       },
       {
-        label: 'Req. Amount',
-        name: 'requested_amount',
+        label: 'Loan Amount',
+        name: 'new_loan_amount',
         options: {
           filter: false,
           sort: true,
           setCellProps: () => ({
-            align: 'left',
+            align: 'right',
+          }),
+          setCellHeaderProps: () => ({
+            align: 'right',
           }),
           customBodyRender: value => <strong><Currency value={value} /></strong>
-        }
-      },
-      {
-        label: 'Month of renewal',
-        name: 'renewal_month',
-        options: {
-          filter: true,
-          filterWidth: '100%',
-          sort: true,
-          setCellProps: () => ({
-            align: 'center',
-          }),
-          customBodyRender: value => {
-            return <div>{value ? moment(new Date(value), 'YYYY-MM-DD').format('MMM, YY') : '-'}</div>
-          }
-        }
-      },
-      {
-        label: 'Renewal Fee status',
-        name: 'renewal_fee_payment_status',
-        options: {
-          filter: false,
-          sort: true,
-          customBodyRender: (value) => {
-            return <>{value?.toUpperCase()}</>
-          },
         }
       },
     ]
@@ -154,30 +135,11 @@ const ReviewTable = ({ title, onRowClick, filterQry, currentUser }) => {
     filter: false,
     print: false,
     sort: false,
+    download: false,
     viewColumns: false,
     searchPlaceholder: 'Search by dealreship ID/Name',
     onSearchChange: (searchText) => {
       setSearch(searchText)
-    },
-    customFooter: () => {
-      return (
-        <div style={{ display: 'flex', justifyContent: 'flex-end' }}>
-          <MuiTableFooter
-            totalCount={pageDetailsQuery?.data?.total_number_of_pages}
-            pageSize={10}
-            onPageChange={(value) => { setPage(value) }}
-          />
-        </div>
-      )
-    },
-    onCellClick: (colData, cellMeta) => {
-      if (cellMeta.colIndex !== 7) {
-        onRowClick(loans[cellMeta.dataIndex].dealership_id, loans[cellMeta.dataIndex], 'approval')
-      }
-    },
-    customSort: (data, dataIndex, rowIndex) => {
-      let dateIndex = 5
-      return dateCustomSort(data, dataIndex, rowIndex, dateIndex)
     },
     customToolbar: () => {
       return (
@@ -188,13 +150,34 @@ const ReviewTable = ({ title, onRowClick, filterQry, currentUser }) => {
         </>
       );
     },
+    customFooter: () => {
+      return (
+        <div style={{ display: 'flex', justifyContent: 'flex-end' }}>
+          <MuiTableFooter
+            totalCount={pageData?.total_number_of_pages}
+            pageSize={10}
+            onPageChange={(value) => { setPage(value) }}
+          />
+        </div>
+      )
+    },
+    onCellClick: (colData, cellMeta) => {
+      if (cellMeta.colIndex !== 7) {
+        onRowClick(loans[cellMeta.dataIndex].dealership_id, loans[cellMeta.dataIndex], 'submit')
+      }
+    },
+    customSort: (data, dataIndex, rowIndex) => {
+      let dateIndex = 5
+      return dateCustomSort(data, dataIndex, rowIndex, dateIndex)
+    }
   };
 
   return (
     <div className={classes.root}>
       <MUIDataTable
-        title={title ? <Typography className={classes.title} variant="h4" component="h4">{title}</Typography> : null}
+        title={title ? <Typography className={classes.title} variant="h4" component="h4">{title} ({loans.length})</Typography> : null}
         data={loans}
+        style={classes.tableStyle}
         columns={columns}
         options={options}
       />
@@ -205,4 +188,4 @@ const ReviewTable = ({ title, onRowClick, filterQry, currentUser }) => {
   )
 }
 
-export default ReviewTable;
+export default SubmittedTable;